+++ algoCpp/source/main/include/nbEnregistrements.hpp
@@ -21,24 +21,24 @@
 
 using namespace std;//ça permet d'utiliser la librairie stl
 
-typedef vector<vector<int>> matrice;//on définit le type matrice
+typedef vector<vector<long>> matrice;//on définit le type matrice
 typedef vector<vector<string>> matrice_string;//on définit le type matrice pour des string
 
 vector<vector<string>> chargerFichiers(string filePath);
 
 matrice conversion(vector<vector<string>> tableFaitString);
 
-void generer(matrice& M, int l, int c);
+void generer(matrice& M, long l, long c);
 
-vector<int> extraire(vector<int>& V, vector<int>& W);
+vector<long> extraire(vector<long>& V, vector<long>& W);
 
-int taille(matrice& M, vector<int>& W);
+long taille(matrice& M, vector<long>& W);
 
-vector<int> convertirEnBinaire(int n, int c);
+vector<long> convertirEnBinaire(long n, long c);
 
-vector<int> Binaire_colonnes(vector<int> &V);
+vector<long> Binaire_colonnes(vector<long> &V);
 
-vector<int> toutes_les_tailles(matrice & M);
+vector<long> toutes_les_tailles(matrice & M);
 
 
 #endif /* nbEnregistrements_hpp */
